requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
python-dotenv==1.2.1
